Исправления замечаний после 6 урока устранены
