+++ lesson5/admin/product-edit.php
@@ -6,10 +6,10 @@
  * 
  */
 
-require('admin-link.php');
-require('updateProduct.php');
+require('./link.php');
+require('functions.php');
 
-$productId = $_GET['productId'];
+$productId = (int) $_GET['productId'];
 
 /**
  * Функция подключается к базе данных, возвращает картинку, инкрементирует количество посещений
@@ -61,7 +61,7 @@ function renderImage($id, $dir)
 	<header>
 		<h2 class="heading">Product</h2>	
 	</header>
-	<nav class="nav"><a class="nav-link" href="catalog-admin.php">Gallery</a></nav>
+	<nav class="nav"><a class="nav-link" href="admin.php">Gallery</a></nav>
 
 	<section class="product-edit">
 			<?php 
@@ -84,7 +84,7 @@ function renderImage($id, $dir)
 					</form>
 				</div>
 				<div class="catalog__delete">
-					<a href="./delProduct.php?productId=<?= $productId ?>"><i class='fas fa-trash-alt catalog__del-big'></i></a>
+					<a href="./delete.php?productId=<?= $productId ?>"><i class='fas fa-trash-alt catalog__del-big'></i></a>
 				</div>
 			</div>
 			<?php
